back-end/requirements.txt
Flask~=3.1.0
Flask-Cors~=5.0.0
Flask-SQLAlchemy==3.0.4
PyJWT~=2.10.1
Werkzeug~=3.1.3
pytz==2024.2
pandas~=2.2.3
sqlalchemy~=2.0.36
flask-migrate~=4.0.7